Modified file

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ long int long_int;
 long long int long2_int;
 float float_size;
 
-printf("Size of a char: %lu byte(s)\n",(unsigned long)sizeof(char_size));
-printf("Size of an int: %lu byte(s)\n",(unsigned long)sizeof(int_size));
-printf("Size of a long int: %lu byte(s)\n",(unsigned long)sizeof(long_int));
-printf("Size of a long long int: %lu byte(s)\n",(unsigned long)sizeof(long2_int));
-printf("Size of a float: %lu byte(s)\n",(unsigned long)sizeof(float_size));
+printf("Size of a char: %d byte(s)\n", (int)sizeof(char_size));
+printf("Size of an int: %d byte(s)\n", (int)sizeof(int_size));
+printf("Size of a long int: %d byte(s)\n", (int)sizeof(long_int));
+printf("Size of a long long int: %d byte(s)\n", (int)sizeof(long2_int));
+printf("Size of a float: %d byte(s)\n", (int)sizeof(float_size));
 
 return (0);
 }
